refactor(authflow): untangle the flow with a single advance dispatcher
The authreq and authcode packages moved code around without clarifying the
flow, so fold them back into the authflow package as files — request.go
(parse/validate the /auth request) and response.go (issue the code/token
response) — the way fosite keeps request handling and response writing as files
on one provider rather than separate packages.

The real fix is the control flow. The next-hop decision (next MFA factor,
consent screen, or issue) was scattered across finalizeLogin, finishSessionLogin
and the approval handler, each building the URL itself with a different return
shape ((url, skip, err) vs (url, bool)). Introduce advance (nextstep.go): the
single place that reads nextAuthStep and dispatches — redirect to MFA or consent,
or issue the code — in the spirit of zitadel's NextSteps. finalizeLogin now just
finalizes the identity and returns the auth request; connector login, callback
and session login each finalize their step and call advance. prompt=none's
"no interaction" rule lives in advance too, so it is enforced in one place.

Conventions: package docs moved to doc.go in every package, matching the rest of
the tree; the test harness file is handler_test.go (not helpers_test.go), and
the session tests are sessionlogin_test.go, mirroring their source.
