+++ server/authflow/mfa/doc.go
@@ -1,0 +1,3 @@
+// Package mfa owns dex's MFA domain: the authenticator chain and the TOTP
+// and WebAuthn endpoints. The interactive auth flow delegates to a Manager.
+package mfa
