+++ server/authflow/nextstep.go
@@ -2,9 +2,42 @@ package authflow
 
 import (
 	"context"
+	"net/http"
 
+	"github.com/dexidp/dex/server/oauth2"
 	"github.com/dexidp/dex/storage"
 )
+
+// advance runs the auth request's next step: it redirects to the next MFA factor
+// or the consent screen, or issues the authorization code. It is the single
+// place the post-identity transition lives — connector login, session login,
+// approval and MFA completion all funnel through it instead of each deciding
+// where to go and building the URL themselves.
+func (h *Handler) advance(w http.ResponseWriter, r *http.Request, authReq storage.AuthRequest) {
+	step, err := h.nextAuthStep(r.Context(), &authReq)
+	if err != nil {
+		h.logger.ErrorContext(r.Context(), "failed to determine next auth step", "err", err)
+		h.RenderError(r, w, http.StatusInternalServerError, ErrMsgInternalServerError)
+		return
+	}
+	// prompt=none forbids any user-facing interaction: if the flow can't complete
+	// silently, report interaction_required instead of showing MFA or consent.
+	if prompt, _ := oauth2.ParsePrompt(authReq.Prompt); prompt.None() {
+		if _, ok := step.(issueStep); !ok {
+			h.redirectWithError(w, r, &authReq, oauth2.InteractionRequired, "User interaction required")
+			return
+		}
+	}
+
+	switch step := step.(type) {
+	case mfaStep:
+		http.Redirect(w, r, h.mfa.BuildRedirectURL(authReq, step.authenticator), http.StatusSeeOther)
+	case approvalStep:
+		http.Redirect(w, r, h.BuildApprovalURL(authReq), http.StatusSeeOther)
+	case issueStep:
+		h.sendCodeResponse(w, r, authReq)
+	}
+}
 
 // authStep is the next thing a logged-in authorization request needs before the
 // code can be issued. It is data, not behavior — the whole "what's next"
