+++ server/authflow/request_test.go
@@ -1,4 +1,4 @@
-package authreq
+package authflow
 
 import (
 	"crypto/rand"
@@ -16,6 +16,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/dexidp/dex/server/authflow/web"
 	"github.com/dexidp/dex/server/oauth2"
 	"github.com/dexidp/dex/server/signer"
 	"github.com/dexidp/dex/server/tokens"
@@ -83,7 +84,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "code",
 				"scope":         "openid email profile",
 			},
-			expectedError: &DisplayedErr{Status: http.StatusNotFound},
+			expectedError: &displayedAuthErr{Status: http.StatusNotFound},
 		},
 		{
 			name: "invalid redirect uri",
@@ -100,7 +101,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "code",
 				"scope":         "openid email profile",
 			},
-			expectedError: &DisplayedErr{Status: http.StatusBadRequest},
+			expectedError: &displayedAuthErr{Status: http.StatusBadRequest},
 		},
 		{
 			name: "implicit flow",
@@ -133,7 +134,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "code id_token",
 				"scope":         "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.UnsupportedResponseType},
+			expectedError: &redirectedAuthErr{Type: oauth2.UnsupportedResponseType},
 		},
 		{
 			name: "only token response type",
@@ -150,7 +151,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "token",
 				"scope":         "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "choose connector_id",
@@ -202,7 +203,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "code id_token",
 				"scope":         "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "PKCE code_challenge_method plain",
@@ -274,7 +275,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"code_challenge_method": "invalid_method",
 				"scope":                 "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "No response type",
@@ -292,7 +293,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"code_challenge_method": "plain",
 				"scope":                 "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "PKCE enforced, no code_challenge provided",
@@ -313,7 +314,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"response_type": "code",
 				"scope":         "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "PKCE enforced, code_challenge provided",
@@ -357,7 +358,7 @@ func TestParseAuthorizationRequest(t *testing.T) {
 				"code_challenge_method": "plain",
 				"scope":                 "openid email profile",
 			},
-			expectedError: &RedirectedErr{Type: oauth2.InvalidRequest},
+			expectedError: &redirectedAuthErr{Type: oauth2.InvalidRequest},
 		},
 		{
 			name: "PKCE only S256 allowed, S256 accepted",
@@ -384,7 +385,14 @@ func TestParseAuthorizationRequest(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			p := newTestParser(t, tc.clients, tc.pkce, toResponseTypeSet(tc.supportedResponseTypes), nil)
+			httpServer, server := newTestHandler(t, func(c *Config) {
+				c.SupportedResponseTypes = toResponseTypeSet(tc.supportedResponseTypes)
+				c.Storage = storage.WithStaticClients(c.Storage, tc.clients)
+				if len(tc.pkce.CodeChallengeMethodsSupported) > 0 || tc.pkce.Enforce {
+					c.PKCE = tc.pkce
+				}
+			})
+			defer httpServer.Close()
 
 			params := url.Values{}
 			for k, v := range tc.queryParams {
@@ -393,23 +401,23 @@ func TestParseAuthorizationRequest(t *testing.T) {
 			var req *http.Request
 			if tc.usePOST {
 				body := strings.NewReader(params.Encode())
-				req = httptest.NewRequest("POST", "/auth", body)
+				req = httptest.NewRequest("POST", httpServer.URL+"/auth", body)
 				req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 			} else {
-				req = httptest.NewRequest("GET", "/auth?"+params.Encode(), nil)
+				req = httptest.NewRequest("GET", httpServer.URL+"/auth?"+params.Encode(), nil)
 			}
 
-			_, _, err := p.Parse(req)
+			_, _, err := server.parseAuthorizationRequest(req)
 			if tc.expectedError == nil {
 				if err != nil {
 					t.Errorf("%s: expected no error", tc.name)
 				}
 			} else {
 				switch expectedErr := tc.expectedError.(type) {
-				case *RedirectedErr:
-					e, ok := err.(*RedirectedErr)
+				case *redirectedAuthErr:
+					e, ok := err.(*redirectedAuthErr)
 					if !ok {
-						t.Fatalf("%s: expected RedirectedErr error", tc.name)
+						t.Fatalf("%s: expected redirectedAuthErr error", tc.name)
 					}
 					if e.Type != expectedErr.Type {
 						t.Errorf("%s: expected error type %v, got %v", tc.name, expectedErr.Type, e.Type)
@@ -417,10 +425,10 @@ func TestParseAuthorizationRequest(t *testing.T) {
 					if e.RedirectURI != tc.queryParams["redirect_uri"] {
 						t.Errorf("%s: expected error to be returned in redirect to %v", tc.name, tc.queryParams["redirect_uri"])
 					}
-				case *DisplayedErr:
-					e, ok := err.(*DisplayedErr)
+				case *displayedAuthErr:
+					e, ok := err.(*displayedAuthErr)
 					if !ok {
-						t.Fatalf("%s: expected DisplayedErr error", tc.name)
+						t.Fatalf("%s: expected displayedAuthErr error", tc.name)
 					}
 					if e.Status != expectedErr.Status {
 						t.Errorf("%s: expected http status %v, got %v", tc.name, expectedErr.Status, e.Status)
@@ -685,7 +693,7 @@ func TestRedirectedAuthErrHandler(t *testing.T) {
 	for _, tc := range tests {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
-			err := &RedirectedErr{
+			err := &redirectedAuthErr{
 				State:       tc.state,
 				RedirectURI: tc.redirectURI,
 				Type:        tc.errType,
@@ -765,7 +773,12 @@ func TestValidateIDTokenHint(t *testing.T) {
 	issuerURL, err := url.Parse("https://issuer.example.com")
 	require.NoError(t, err)
 
-	s := New(nil, slog.Default(), sig, *issuerURL, PKCEConfig{}, nil)
+	s := &Handler{
+		signer:    sig,
+		issuerURL: *issuerURL,
+		logger:    slog.Default(),
+		UI:        web.New(nil, *issuerURL, slog.Default()),
+	}
 
 	now := time.Now()
 
@@ -775,7 +788,7 @@ func TestValidateIDTokenHint(t *testing.T) {
 			Subject: "CgNmb28SA2Jhcg",
 			Expiry:  now.Add(1 * time.Hour).Unix(),
 		})
-		idToken, err := s.ValidateIDTokenHint(t.Context(), token)
+		idToken, err := s.validateIDTokenHint(t.Context(), token)
 		require.NoError(t, err)
 		assert.Equal(t, "CgNmb28SA2Jhcg", idToken.Subject)
 	})
@@ -786,7 +799,7 @@ func TestValidateIDTokenHint(t *testing.T) {
 			Subject: "CgNmb28SA2Jhcg",
 			Expiry:  now.Add(-1 * time.Hour).Unix(),
 		})
-		idToken, err := s.ValidateIDTokenHint(t.Context(), token)
+		idToken, err := s.validateIDTokenHint(t.Context(), token)
 		require.NoError(t, err)
 		assert.Equal(t, "CgNmb28SA2Jhcg", idToken.Subject)
 	})
@@ -809,7 +822,7 @@ func TestValidateIDTokenHint(t *testing.T) {
 		token, err := jws.CompactSerialize()
 		require.NoError(t, err)
 
-		_, err = s.ValidateIDTokenHint(t.Context(), token)
+		_, err = s.validateIDTokenHint(t.Context(), token)
 		assert.Error(t, err)
 	})
 
@@ -819,22 +832,22 @@ func TestValidateIDTokenHint(t *testing.T) {
 			Subject: "CgNmb28SA2Jhcg",
 			Expiry:  now.Add(1 * time.Hour).Unix(),
 		})
-		_, err := s.ValidateIDTokenHint(t.Context(), token)
+		_, err := s.validateIDTokenHint(t.Context(), token)
 		assert.Error(t, err)
 	})
 
 	t.Run("malformed token", func(t *testing.T) {
-		_, err := s.ValidateIDTokenHint(t.Context(), "not-a-valid-jwt")
+		_, err := s.validateIDTokenHint(t.Context(), "not-a-valid-jwt")
 		assert.Error(t, err)
 	})
 }
 
 func TestSessionMatchesHint(t *testing.T) {
 	// tokens.GenSubject("foo", "bar") == "CgNmb28SA2Jhcg" (from TestGetSubject)
-	assert.True(t, SessionMatchesHint(&storage.AuthSession{UserID: "foo", ConnectorID: "bar"}, "CgNmb28SA2Jhcg"))
-	assert.False(t, SessionMatchesHint(&storage.AuthSession{UserID: "other", ConnectorID: "bar"}, "CgNmb28SA2Jhcg"))
-	assert.False(t, SessionMatchesHint(&storage.AuthSession{UserID: "foo", ConnectorID: "other"}, "CgNmb28SA2Jhcg"))
-	assert.False(t, SessionMatchesHint(nil, "CgNmb28SA2Jhcg"))
+	assert.True(t, sessionMatchesHint(&storage.AuthSession{UserID: "foo", ConnectorID: "bar"}, "CgNmb28SA2Jhcg"))
+	assert.False(t, sessionMatchesHint(&storage.AuthSession{UserID: "other", ConnectorID: "bar"}, "CgNmb28SA2Jhcg"))
+	assert.False(t, sessionMatchesHint(&storage.AuthSession{UserID: "foo", ConnectorID: "other"}, "CgNmb28SA2Jhcg"))
+	assert.False(t, sessionMatchesHint(nil, "CgNmb28SA2Jhcg"))
 }
 
 func TestParseAuthorizationRequest_IDTokenHint(t *testing.T) {
@@ -844,10 +857,17 @@ func TestParseAuthorizationRequest_IDTokenHint(t *testing.T) {
 	now := time.Now()
 
 	t.Run("valid id_token_hint populates subject", func(t *testing.T) {
-		p := newTestParser(t, []storage.Client{{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}}}, PKCEConfig{}, map[string]bool{"code": true}, sig)
+		httpServer, server := newTestHandler(t, func(c *Config) {
+			c.SupportedResponseTypes = map[string]bool{"code": true}
+			c.Storage = storage.WithStaticClients(c.Storage, []storage.Client{
+				{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}},
+			})
+			c.Signer = sig
+		})
+		defer httpServer.Close()
 
 		token := signTestIDToken(t, tokens.IDTokenClaims{
-			Issuer:  testIssuer,
+			Issuer:  httpServer.URL,
 			Subject: "CgNmb28SA2Jhcg",
 			Expiry:  now.Add(1 * time.Hour).Unix(),
 		})
@@ -859,15 +879,22 @@ func TestParseAuthorizationRequest_IDTokenHint(t *testing.T) {
 			"scope":         {"openid"},
 			"id_token_hint": {token},
 		}
-		req := httptest.NewRequest("GET", "/auth?"+params.Encode(), nil)
+		req := httptest.NewRequest("GET", httpServer.URL+"/auth?"+params.Encode(), nil)
 
-		_, hintSubject, err := p.Parse(req)
+		_, hintSubject, err := server.parseAuthorizationRequest(req)
 		require.NoError(t, err)
 		assert.Equal(t, "CgNmb28SA2Jhcg", hintSubject)
 	})
 
 	t.Run("invalid id_token_hint returns error", func(t *testing.T) {
-		p := newTestParser(t, []storage.Client{{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}}}, PKCEConfig{}, map[string]bool{"code": true}, sig)
+		httpServer, server := newTestHandler(t, func(c *Config) {
+			c.SupportedResponseTypes = map[string]bool{"code": true}
+			c.Storage = storage.WithStaticClients(c.Storage, []storage.Client{
+				{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}},
+			})
+			c.Signer = sig
+		})
+		defer httpServer.Close()
 
 		params := url.Values{
 			"client_id":     {"foo"},
@@ -876,17 +903,23 @@ func TestParseAuthorizationRequest_IDTokenHint(t *testing.T) {
 			"scope":         {"openid"},
 			"id_token_hint": {"invalid-token"},
 		}
-		req := httptest.NewRequest("GET", "/auth?"+params.Encode(), nil)
+		req := httptest.NewRequest("GET", httpServer.URL+"/auth?"+params.Encode(), nil)
 
-		_, _, err := p.Parse(req)
+		_, _, err := server.parseAuthorizationRequest(req)
 		require.Error(t, err)
-		redirectErr, ok := err.(*RedirectedErr)
+		redirectErr, ok := err.(*redirectedAuthErr)
 		require.True(t, ok)
 		assert.Equal(t, oauth2.InvalidRequest, redirectErr.Type)
 	})
 
 	t.Run("no id_token_hint leaves subject empty", func(t *testing.T) {
-		p := newTestParser(t, []storage.Client{{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}}}, PKCEConfig{}, map[string]bool{"code": true}, nil)
+		httpServer, server := newTestHandler(t, func(c *Config) {
+			c.SupportedResponseTypes = map[string]bool{"code": true}
+			c.Storage = storage.WithStaticClients(c.Storage, []storage.Client{
+				{ID: "foo", RedirectURIs: []string{"https://example.com/foo"}},
+			})
+		})
+		defer httpServer.Close()
 
 		params := url.Values{
 			"client_id":     {"foo"},
@@ -894,9 +927,9 @@ func TestParseAuthorizationRequest_IDTokenHint(t *testing.T) {
 			"response_type": {"code"},
 			"scope":         {"openid"},
 		}
-		req := httptest.NewRequest("GET", "/auth?"+params.Encode(), nil)
+		req := httptest.NewRequest("GET", httpServer.URL+"/auth?"+params.Encode(), nil)
 
-		_, hintSubject, err := p.Parse(req)
+		_, hintSubject, err := server.parseAuthorizationRequest(req)
 		require.NoError(t, err)
 		assert.Equal(t, "", hintSubject)
 	})
